+++ config/react.php
@@ -1,0 +1,27 @@
+<?php
+
+return [
+    /*
+    |--------------------------------------------------------------------------
+    | Path Prefix
+    |--------------------------------------------------------------------------
+    |
+    | Change the default path prefix under resources/js used by the command
+    | when a relative path is provided as the name argument.
+    |
+    */
+    'prefix' => 'components',
+
+    /*
+    |--------------------------------------------------------------------------
+    | Default Options
+    |--------------------------------------------------------------------------
+    |
+    | Provide default options to be passed to the command when it is called.
+    | For example, to always use TypeScript and the file extension .ts
+    | then change the value to ['typescript' => true, 'extension' => 'ts'].
+    | Options should be in the same format accepted by Artisan::call().
+    |
+    */
+    'defaults' => [],
+];