+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release Notes
 
-## Upcoming
+## v4.0.0
 - Dropped support for Laravel 9
 - Dropped support for PHP 8.0 
 - Removed support for prop-types